Conflicts:
	src/main/java/org/basex/query/QueryContext.java
	src/main/java/org/basex/query/QueryParser.java
	src/main/java/org/basex/query/expr/Catch.java
	src/main/java/org/basex/query/expr/CmpG.java
	src/main/java/org/basex/query/expr/List.java
	src/main/java/org/basex/query/expr/Preds.java
	src/main/java/org/basex/query/expr/Quantifier.java
	src/main/java/org/basex/query/expr/VarRef.java
	src/main/java/org/basex/query/flwor/GroupPartition.java
	src/main/java/org/basex/query/func/DynamicFunc.java
	src/main/java/org/basex/query/func/FNFt.java
	src/main/java/org/basex/query/func/Function.java
	src/main/java/org/basex/query/func/Functions.java
	src/main/java/org/basex/query/func/InlineFunc.java
	src/main/java/org/basex/query/func/PartFunc.java
	src/main/java/org/basex/query/func/StandardFunc.java
	src/main/java/org/basex/query/func/UserFuncs.java
	src/main/java/org/basex/query/path/MixedPath.java
	src/main/java/org/basex/query/up/expr/Transform.java
	src/main/java/org/basex/query/util/Err.java
	src/main/java/org/basex/query/util/Var.java
	src/test/java/org/basex/test/query/ast/FlworOptimizeTest.java

+++ src/main/java/org/basex/index/path/PathSummary.java
@@ -191,8 +191,7 @@ public final class PathSummary implements Index {
     for(final byte[] i : tl) {
       final boolean att = startsWith(i, '@');
       final byte kind = att ? Data.ATTR : Data.ELEM;
-      final int id = att ? data.atnindex.id(substring(i, 1)) :
-        data.tagindex.id(i);
+      final int id = att ? data.atnindex.id(substring(i, 1)) : data.tagindex.id(i);
 
       final ArrayList<PathNode> out = new ArrayList<PathNode>();
       for(final PathNode n : in) {
@@ -206,7 +205,7 @@ public final class PathSummary implements Index {
     }
 
     // sort by number of occurrences
-    final double[] tmp = new double[in.size()];
+    final int[] tmp = new int[in.size()];
     for(int i = 0; i < in.size(); ++i) tmp[i] = in.get(i).stats.count;
     final int[] occ = Array.createOrder(tmp, false);
